+++ python/ray/rllib/__init__.py
@@ -8,16 +8,16 @@ import logging
 # This file is imported from the tune module in order to register RLlib agents.
 from ray.tune.registry import register_trainable
 
-from ray.rllib.policy.policy import Policy
-from ray.rllib.policy.policy_graph import PolicyGraph
-from ray.rllib.policy.tf_policy import TFPolicy
-from ray.rllib.policy.tf_policy_graph import TFPolicyGraph
+from ray.rllib.evaluation.policy_graph import PolicyGraph
+from ray.rllib.evaluation.tf_policy_graph import TFPolicyGraph
+from ray.rllib.evaluation.policy_evaluator import PolicyEvaluator
 from ray.rllib.env.base_env import BaseEnv
 from ray.rllib.env.multi_agent_env import MultiAgentEnv
 from ray.rllib.env.vector_env import VectorEnv
 from ray.rllib.env.external_env import ExternalEnv
-from ray.rllib.policy.policy_evaluator import PolicyEvaluator
-from ray.rllib.evaluation.sample_batch import SampleBatch
+from ray.rllib.policy.policy import Policy
+from ray.rllib.policy.tf_policy import TFPolicy
+from ray.rllib.policy.sample_batch import SampleBatch
 
 
 def _setup_logger():
