+++ python/ray/rllib/agents/pg/torch_pg_policy.py
@@ -5,8 +5,8 @@ from __future__ import print_function
 import ray
 from ray.rllib.evaluation.postprocessing import compute_advantages, \
     Postprocessing
-from ray.rllib.evaluation.sample_batch import SampleBatch
-from ray.rllib.evaluation.torch_policy_template import build_torch_policy
+from ray.rllib.policy.sample_batch import SampleBatch
+from ray.rllib.policy.torch_policy_template import build_torch_policy
 
 
 def pg_torch_loss(policy, batch_tensors):
