+++ python/ray/rllib/evaluation/__init__.py
@@ -1,11 +1,8 @@
 from ray.rllib.evaluation.episode import MultiAgentEpisode
-from ray.rllib.policy.policy_evaluator import PolicyEvaluator
+from ray.rllib.evaluation.policy_evaluator import PolicyEvaluator
 from ray.rllib.evaluation.interface import EvaluatorInterface
-from ray.rllib.policy.policy import Policy
-from ray.rllib.policy.policy_graph import PolicyGraph
-from ray.rllib.policy.tf_policy import TFPolicy
-from ray.rllib.policy.tf_policy_graph import TFPolicyGraph
-from ray.rllib.evaluation.torch_policy import TorchPolicy
+from ray.rllib.evaluation.policy_graph import PolicyGraph
+from ray.rllib.evaluation.tf_policy_graph import TFPolicyGraph
 from ray.rllib.evaluation.torch_policy_graph import TorchPolicyGraph
 from ray.rllib.evaluation.sample_batch import SampleBatch, MultiAgentBatch
 from ray.rllib.evaluation.sample_batch_builder import (
@@ -15,8 +12,8 @@ from ray.rllib.evaluation.postprocessing import compute_advantages
 from ray.rllib.evaluation.metrics import collect_metrics
 
 __all__ = [
-    "EvaluatorInterface", "PolicyEvaluator", "Policy", "TFPolicy",
-    "TorchPolicy", "PolicyGraph", "TFPolicyGraph", "TorchPolicyGraph",
+    "EvaluatorInterface", "PolicyEvaluator",
+    "PolicyGraph", "TFPolicyGraph", "TorchPolicyGraph",
     "SampleBatch", "MultiAgentBatch", "SampleBatchBuilder",
     "MultiAgentSampleBatchBuilder", "SyncSampler", "AsyncSampler",
     "compute_advantages", "collect_metrics", "MultiAgentEpisode"
