+++ python/ray/rllib/policy/policy.py
@@ -7,6 +7,10 @@ import gym
 
 from ray.rllib.utils.annotations import DeveloperAPI
 
+# By convention, metrics from optimizing the loss can be reported in the
+# `grad_info` dict returned by learn_on_batch() / compute_grads() via this key.
+LEARNER_STATS_KEY = "learner_stats"
+
 
 @DeveloperAPI
 class Policy(object):
