+++ python/ray/rllib/policy/torch_policy.py
@@ -12,8 +12,7 @@ try:
 except ImportError:
     pass  # soft dep
 
-from ray.rllib.evaluation.metrics import LEARNER_STATS_KEY
-from ray.rllib.policy.policy import Policy
+from ray.rllib.policy.policy import Policy, LEARNER_STATS_KEY
 from ray.rllib.utils.annotations import override
 from ray.rllib.utils.tracking_dict import UsageTrackingDict
