+++ python/ray/rllib/policy/torch_policy_template.py
@@ -3,7 +3,7 @@ from __future__ import division
 from __future__ import print_function
 
 from ray.rllib.policy.policy import Policy
-from ray.rllib.evaluation.torch_policy import TorchPolicy
+from ray.rllib.policy.torch_policy import TorchPolicy
 from ray.rllib.models.catalog import ModelCatalog
 from ray.rllib.utils.annotations import override, DeveloperAPI
